ZMQx::Class receive_string methods
The Python lib for zmq implements receive_string for obvious reasons. I like
the approach and have added the receive_string() to ZMQx::Class::Socket. It is
possible to use encodings other than 'utf8' but it will be slower.

beware!! we're using Encode::_utf8_on for 'upgrading' the strings. This is
evil (but fast).
 (RT63420_ZMQx_Class_Unicode_safe)