Update send() to reflect the (non-)return value from ZMQ::FFI::send().
ZMQ::LibZMQ3() is documented to return the number of bytes sent (which is the
return value from the underlying zmq_msg_send() C library call).
ZMQ::FFI::send() is not documented to return anything in particular, and will
throw an exception if the C library call returns -1 for error.

The code in ZMQx::Class was relying on the ZMQ::LibZMQ3 behaviour, and getting
caught out if ZMQ::FFI::send() happened to return a false value.

For now, ignore the return value of ZMQ::FFI::send() and send_multipart(), and
continue to honour our documented return value (the length of the message) by
using the value passed in.

It's unclear if this is the best long term fix.